+++ CommandArguments.h
@@ -14,7 +14,7 @@ class CommandArguments {
 public:
   typedef CommandOptionStorage<std::string> StringStorageType;
   typedef CommandOptionStorage<bool> FlagStorageType;
-  typedef CommandOptionStorage<__int64> NumStorageType;
+  typedef CommandOptionStorage<long long> NumStorageType;
 
   typedef CommandOptionStorage<std::vector<std::string>> StringListStorageType;
 
@@ -31,7 +31,7 @@ public:
   StringStorageType *AddParam(const std::string &name, const std::string &desc,
                               const std::string &default_value);
   NumStorageType *AddNumber(const std::string &name, const std::string &desc,
-                            __int64 default_value);
+                            long long default_value);
   FlagStorageType *AddFlag(const std::string &name, const std::string &desc,
                            bool default_value);
   StringListStorageType *AddParamList(const std::string &name,
